상품 갯수 증감소

+++ MaskProject/src/co/mask/product/web/DeleteAdminExec.java
@@ -1,11 +1,14 @@
 package co.mask.product.web;
 
+
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import co.mask.cart.dao.CartDao;
+import co.mask.cart.vo.CartVo;
 import co.mask.common.Command;
-import co.mask.member.vo.MemberVo;
 import co.mask.product.dao.ProductDao;
 import co.mask.product.vo.ProductVo;
 
@@ -16,15 +19,21 @@ public class DeleteAdminExec implements Command {
 		// DeleteAdminExec.do 처리 메소드
 		ProductDao dao = new ProductDao();
 		ProductVo vo = new ProductVo();
-		MemberVo Mvo = new MemberVo();
+		
 		HttpSession session = request.getSession();
+		Object cartview =  session.getAttribute("cartProduct");
+		String cview = (String)cartview;
 		
-		String admin1=(String) session.getAttribute("memberAuth");
-		System.out.println(admin1);
+		System.out.println(cview);
 		
-		Mvo.setMemberAuth(admin1);
-		
-		int n=dao.deleteAdmin(vo);
+		if(cartview != null) {
+			request.setAttribute("msg", "해당상품은 출고상품으로 삭제 불가능합니다.");
+		}else {
+			vo.setProductNum(Integer.parseInt(request.getParameter("row")));
+			int n=dao.deleteAdmin(vo);
+			System.out.println(n);
+			request.setAttribute("vo", vo);
+		}
 		
 		return "showProduct.do";
 	}
